% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_stan_probability_distributions}
\alias{available_stan_probability_distributions}
\title{Available Stan Probability Distributions}
\format{
A data frame with 1 row and 5 variables:
\describe{
\item{function_name}{Name of the function}
}
}
\usage{
available_stan_probability_distributions
}
\description{
A data.frame listing all the Stan probability distribution functions
available within this package.
}
\details{
The source-code for a function can be loaded using the
\code{stan_probability_distribution} function
}
\examples{
available_stan_probability_distributions
}
\keyword{datasets}
