% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_lengths}
\alias{check_lengths}
\title{compare arguments lengths, probably smarter way of doing this than combn}
\usage{
check_lengths(...)
}
\arguments{
\item{...}{objects to compare lengths of}
}
\description{
compare arguments lengths, probably smarter way of doing this than combn
}
