% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomial.R
\name{dnbinom_stan}
\alias{dnbinom_stan}
\title{Stan code for the Negative Binomial Log Probability Density}
\usage{
dnbinom_stan()
}
\value{
A text string containing Stan code.
}
\description{
Stan has built-in functions for calculating the log probability density
of the negative binomial distribution:
\preformatted{
real neg_binomial_lpmf(ints n | reals alpha, reals beta)
real neg_binomial_2_lpmf(ints y | reals mu, reals phi)
real neg_binomial_2_log_lpmf(ints y | reals eta, reals phi)
}
}
