% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomial.R
\name{dneg_binomial_2}
\alias{dneg_binomial_2}
\title{The Negative Binomial Probability Mass Function (alternative parameterisation)}
\usage{
dneg_binomial_2(x, mu, phi, log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles.}

\item{mu}{mean.}

\item{phi}{dispersion.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
the (log) probability mass at x, given mu and phi.
}
\description{
The Negative Binomial Probability Mass Function (alternative parameterisation)
}
