% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{dpois_stan}
\alias{dpois_stan}
\title{Stan code for the Poisson Log Probability Density}
\usage{
dpois_stan()
}
\value{
A text string containing Stan code.
}
\description{
Stan already has built-in support for the Poisson distribution via the
poisson_lpmf function.
Therefore, this function just returns some comments directing people
to that.
}
