% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailAdjustedBeta.R
\name{dtab}
\alias{dtab}
\title{Tail-adjusted Beta Probability Density Function}
\usage{
dtab(x, mu, phi, delta, log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative) values.}

\item{mu}{mean/s in (0, 1) for the beta distribution.}

\item{phi}{dispersion parameter (positive) for the beta distribution.}

\item{delta}{rounding parameter in (0, 0.5) for the tail adjustment.}

\item{log}{logical; if TRUE, probabilities, p, are given as log(p).}
}
\value{
The (log) probability mass at x, given lambda and pi.
}
\description{
Tail-adjusted Beta Probability Density Function
}
