% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroAugmentedGamma.R
\name{dzag}
\alias{dzag}
\title{Zero-augmented Gamma Probability Density Function}
\usage{
dzag(x, mu, phi, pi, log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative) values.}

\item{mu}{mean/s (non-negative) for the gamma distribution.}

\item{phi}{dispersion parameter for the gamma distribution.}

\item{pi}{vector of (real lying in [0, 1]) zero-inflation parameters.}

\item{log}{logical; if TRUE, probabilities, p, are given as log(p).}
}
\value{
The (log) probability mass at x, given lambda and pi.
}
\description{
Zero-augmented Gamma Probability Density Function
}
