% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedNegativeBinomial.R
\name{dzinb_2}
\alias{dzinb_2}
\title{Zero-inflated Negative Binomial Probability Mass}
\usage{
dzinb_2(x, mu, phi, pi, log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative integer) values.}

\item{mu}{mean of the distribution.}

\item{phi}{dispersion parameter.}

\item{pi}{vector of (real lying in [0, 1]) zero-inflation parameters.}

\item{log}{logical; if TRUE, probabilities, p, are given as log(p).}
}
\value{
The (log) probability mass at x, given mu, phi and pi.
}
\description{
Zero-inflated Negative Binomial Probability Mass
}
