% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedNegativeBinomial.R
\name{dzinbinom}
\alias{dzinbinom}
\title{Zero-inflated Negative Binomial Probability Mass}
\usage{
dzinbinom(x, size, prob, mu, log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative integer) values.}

\item{log}{logical; if TRUE, probabilities, p, are given as log(p).}

\item{phi}{vector of over-dispersion values.}

\item{pi}{vector of (real lying in [0, 1]) zero-inflation parameters.}
}
\value{
The (log) probability mass at x, given lambda and pi.
}
\description{
Zero-inflated Negative Binomial Probability Mass
}
