% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedNegativeBinomialLinked.R
\name{dzinbl}
\alias{dzinbl}
\title{Zero-inflated Negative Binomial Linked Probability Mass}
\usage{
dzinbl(x, size, prob, mu, gamma0, gamma1, log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative integer) values.}

\item{size}{target for number of successful trials, or dispersion parameter (the shape parameter of the gamma mixing distribution). Must be strictly positive, need not be integer.}

\item{prob}{probability of success in each trial. 0 < prob <= 1.}

\item{mu}{alternative parametrization via mean: see ‘Details’.}

\item{gamma0}{The zero-inflation to mean link intercept.}

\item{gamma1}{The zero-inflation to mean link gradient.}

\item{log}{logical; if TRUE, probabilities, p, are given as log(p).}
}
\value{
The (log) probability mass at x, given lambda and pi.
}
\description{
Zero-inflated Negative Binomial Linked Probability Mass
}
