% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedNegativeBinomialLinked.R
\name{dzinbl_2}
\alias{dzinbl_2}
\title{Zero-inflated Negative Binomial Linked Probability Mass}
\usage{
dzinbl_2(x, mu, phi, gamma0, gamma1, log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative integer) values.}

\item{mu}{mean of the distribution.}

\item{phi}{dispersion parameter.}

\item{gamma0}{The zero-inflation to mean link intercept.}

\item{gamma1}{The zero-inflation to mean link gradient.}

\item{log}{logical; if TRUE, probabilities, p, are given as log(p).}
}
\value{
The (log) probability mass at x, given lambda and pi.
}
\description{
Zero-inflated Negative Binomial Linked Probability Mass
}
