% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflationPoisson.R
\name{dzip_stan}
\alias{dzip_stan}
\title{Zero-inflated Poisson Log Probability Mass Stan Code}
\usage{
dzip_stan()
}
\value{
A string containing Stan code.
}
\description{
This function returns a string of Stan code that defines the function
\code{real zi_poisson_lmpf(int n, real lambda, real pi)}. When called
the Stan function will return the probability mass at n, for a
zero-inflated Poisson distribution with mean lambda and zero-inflation
parameter pi.
}
\details{
In order to use it the code should be placed with the functions { ... }
block of a Stan model specification and then referenced within
model { ... } block.
}
