% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedPoissonLinked.R
\name{dzipl}
\alias{dzipl}
\title{Zero-inflated Poisson Linked Probability Mass Function}
\usage{
dzipl(x, lambda, gamma0, gamma1, log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative integer) values.}

\item{lambda}{vector of (non-negative) means.}

\item{gamma0}{The zero-inflation to mean link intercept.}

\item{gamma1}{The zero-inflation to mean link gradient.}

\item{log}{logical; if TRUE, probabilities, p, are given as log(p).}

\item{log.link}{logical; if TRUE, then the zero-inflation is linked to the logarithm of the mean.}
}
\value{
The (log) probability mass at x, given lambda and pi.
}
\description{
Zero-inflated Poisson Linked Probability Mass Function
}
