% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedTailAdjustedBeta.R
\name{dzitab}
\alias{dzitab}
\title{Zero-inflated Tail-adjusted Beta Probability Density Function}
\usage{
dzitab(x, mu, phi, delta, pi, log = FALSE)
}
\arguments{
\item{x}{vector of (non-negative) values.}

\item{mu}{mean/s in (0, 1) for the beta distribution.}

\item{phi}{dispersion parameter (positive) for the beta distribution.}

\item{delta}{rounding parameter in (0, 0.5) for the tail adjustment.}

\item{pi}{vector of (real lying in [0, 1]) zero-inflation parameters.}

\item{log}{logical; if TRUE, probabilities, p, are given as log(p).}
}
\value{
The (log) probability mass at x, given mu, phi, delta and pi.
}
\description{
Zero-inflated Tail-adjusted Beta Probability Density Function
}
