% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroAugmentedGamma.R
\name{gamma_mu_to_rate}
\alias{gamma_mu_to_rate}
\alias{gamma_rate_to_mu}
\title{Convert Gamma parameters between mean-dispersion and shape-rate}
\usage{
gamma_mu_to_rate(mu, phi)

gamma_rate_to_mu(shape, rate)
}
\arguments{
\item{mu}{mean/s (non-negative).}

\item{phi}{dispersion parameter.}

\item{shape}{shape parameter.}

\item{rate}{rate/s (or inverse scale/s).

See \link[stats]{dgamma} for more details on shape and rate.}
}
\value{
a list with the converted Gamma parameters
}
\description{
Convert Gamma parameters between mean-dispersion and shape-rate
}
