% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomial.R
\docType{data}
\name{neg_binomial_2_lpmf_stan}
\alias{neg_binomial_2_lpmf_stan}
\title{Negative Binomial Log Probability Mass Function Stan Code}
\description{
There are two negative binomial density functions built into
Stan:
\preformatted{
real neg_binomial_lpmf(ints n | reals alpha, reals beta)
real neg_binomial_2_lpmf(ints y | reals mu, reals phi)
}

The second one uses the mean and over-dispersion parameterisation familiar
to ecologists.

See \url{https://mc-stan.org/docs/2_19/functions-reference/negative-binomial-distribution.html}
and \url{https://mc-stan.org/docs/2_19/functions-reference/nbalt.html} for
more details.
}
\examples{
cat(neg_binomial_2_lpmf_stan[["source_code"]])
}
\keyword{datasets}
