% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_arg}
\alias{pad_arg}
\title{recycle length one vectors to max output length N}
\usage{
pad_arg(arg, N)
}
\arguments{
\item{arg}{object to pad if necessary}

\item{N}{length to pad to}
}
\description{
recycle length one vectors to max output length N
}
