% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\docType{data}
\name{poisson_lpmf_stan}
\alias{poisson_lpmf_stan}
\title{Poisson Log Probability Mass Stan Code}
\format{
An object of class \code{list} of length 1.
}
\usage{
poisson_lpmf_stan
}
\description{
Stan has a built-in function which calculates the log probability density
of the the Poisson distribution:
\preformatted{
real poisson_lpmf(ints n | reals lambda)
}
}
\keyword{datasets}
