% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilityDistributions-package.R
\docType{package}
\name{probabilityDistributions-package}
\alias{probabilityDistributions}
\alias{probabilityDistributions-package}
\title{probabilityDistributions: Useful probability distributions}
\description{
Useful probability distributions with matching Stan code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://primer-e.github.io/probabilityDistributions}
  \item \url{https://github.com/PRIMER-e/probabilityDistributions}
}

}
\author{
\strong{Maintainer}: Andrew Punnett \email{andrew@primer-e.co.nz} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
