% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeBinomial.R
\name{qneg_binomial_2}
\alias{qneg_binomial_2}
\title{The Negative Binomial Quantile Function (alternative parameterisation)}
\usage{
qneg_binomial_2(p, mu, phi, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{mu}{mean.}

\item{phi}{dispersion.}

\item{lower.tail}{logical; if TRUE (default), probablities are \eqn{P[X <= x]}, otherwise, \eqn{[X > x]}.}

\item{log.p}{logical; if TRUE, probabilities are given as log(p).}
}
\value{
A vector of quantiles, each of which correspond to a probability in p.
}
\description{
The Negative Binomial Quantile Function (alternative parameterisation)
}
