% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedNegativeBinomial.R
\name{qzinb_2}
\alias{qzinb_2}
\title{Zero-inflated Negative Binomial Quantile Function}
\usage{
qzinb_2(p, mu, phi, pi, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of quantiles.}

\item{mu}{mean of the distribution.}

\item{phi}{dispersion parameter.}

\item{pi}{vector of (in \eqn{[0, 1]}) zero-inflation parameters.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X <= x]}, otherwise, \eqn{P[X > x]}.}

\item{log.p}{logical; if TRUE, probabilities p. This doesn't affect pi.}
}
\value{
A vector of quantiles, each of which coincide with the respective probability in p.
}
\description{
Zero-inflated Negative Binomial Quantile Function
}
