% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedNegativeBinomialLogLinked.R
\name{qzinbll}
\alias{qzinbll}
\title{Zero-inflated Negative Log-linked Binomial Quantile Function}
\usage{
qzinbll(p, size, prob, mu, gamma0, gamma1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of quantiles.}

\item{size}{target for number of successful trials, or dispersion parameter (the shape parameter of the gamma mixing distribution). Must be strictly positive, need not be integer.}

\item{prob}{probability of success in each trial. 0 < prob <= 1.}

\item{mu}{alternative parametrization via mean: see ‘Details’.}

\item{gamma0}{The zero-inflation to mean link intercept.}

\item{gamma1}{The zero-inflation to mean link gradient.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X <= x]}, otherwise, \eqn{P[X > x]}.}

\item{log.p}{logical; if TRUE, probabilities p. This doesn't affect pi.}
}
\value{
A vector of quantiles, each of which coincide with the respective probability in p.
}
\description{
Zero-inflated Negative Log-linked Binomial Quantile Function
}
