% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedNegativeBinomialLogLinked.R
\name{qzinbll_2}
\alias{qzinbll_2}
\title{Zero-inflated Negative Log-linked Binomial Quantile Function}
\usage{
qzinbll_2(p, mu, phi, gamma0, gamma1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of quantiles.}

\item{mu}{mean of the distribution.}

\item{phi}{dispersion parameter.}

\item{gamma0}{The zero-inflation to mean link intercept.}

\item{gamma1}{The zero-inflation to mean link gradient.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X <= x]}, otherwise, \eqn{P[X > x]}.}

\item{log.p}{logical; if TRUE, probabilities p. This doesn't affect pi.}
}
\value{
A vector of quantiles, each of which coincide with the respective probability in p.
}
\description{
Zero-inflated Negative Log-linked Binomial Quantile Function
}
