% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedPoissonLogLinked.R
\name{qzipll}
\alias{qzipll}
\title{Zero-inflated Poisson Linked Quantile Function}
\usage{
qzipll(p, lambda, gamma0, gamma1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of quantiles.}

\item{lambda}{vector of (non-negative) means.}

\item{gamma0}{The zero-inflation to mean link intercept.}

\item{gamma1}{The zero-inflation to mean link gradient.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X <= x]}, otherwise, \eqn{P[X > x]}.}

\item{log.p}{logical; if TRUE, probabilities p. This doesn't affect pi.}
}
\value{
A vector of quantiles, each of which coincide with the respective probability in p.
}
\description{
Zero-inflated Poisson Linked Quantile Function
}
