% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_probability_distribution.R
\name{stan_probability_distribution}
\alias{stan_probability_distribution}
\title{Find a Stan Function Definition}
\usage{
stan_probability_distribution(function_name)
}
\arguments{
\item{function_name}{Name of the function.}
}
\value{
A string containing the Stan function definition.
}
\description{
Find a Stan Function Definition
}
\examples{
x <- stan_probability_distribution("zi_poisson_lpmf")
cat(x)
}
