% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedNegativeBinomial.R
\docType{data}
\name{zi_neg_binomial_2_lpmf_stan}
\alias{zi_neg_binomial_2_lpmf_stan}
\title{Zero-inflated Negative-binomial Log Probability Mass Function Stan Code}
\description{
Stan code for the zero-inflated negative-binomial log probability mass function.
}
\examples{
cat(zi_neg_binomial_2_lpmf_stan[["source_code"]])
}
\keyword{datasets}
