% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedNegativeBinomialLinked.R
\docType{data}
\name{zi_neg_binomial_linked_2_lpmf_stan}
\alias{zi_neg_binomial_linked_2_lpmf_stan}
\title{Zero-inflated Negative-binomial Linked Log Probability Mass Function Stan Code}
\description{
Stan code for the zero-inflated negative-binomial distribution's log probability mass function,
where the amount of zero-inflation is a function of the mean.
}
\examples{
cat(zi_neg_binomial_linked_2_lpmf_stan[["source_code"]])
}
\keyword{datasets}
