% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedNegativeBinomialLogLinked.R
\docType{data}
\name{zi_neg_binomial_log_linked_2_lpmf_stan}
\alias{zi_neg_binomial_log_linked_2_lpmf_stan}
\title{Zero-inflated Negative-binomial Log-linked Log Probability Mass Function Stan Code}
\description{
Stan code for the zero-inflated negative-binomial log probability mass function,
where the amount of zero-inflation is a function of log(mean).
}
\examples{
cat(zi_neg_binomial_log_linked_2_lpmf_stan[["source_code"]])
}
\keyword{datasets}
