% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedPoisson.R
\docType{data}
\name{zi_poisson_lpmf_stan}
\alias{zi_poisson_lpmf_stan}
\title{Zero-inflated Poisson Log Probability Mass Function Stan Code}
\description{
Stan code for the zero-inflated Poisson distribution's probability mass function.
}
\examples{
cat(zi_poisson_lpmf_stan[["source_code"]])
}
\keyword{datasets}
