% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_probability_distribution.R
\name{available_stan_probability_distributions}
\alias{available_stan_probability_distributions}
\title{Available Stan Function Definitions}
\usage{
available_stan_probability_distributions()
}
\value{
A dataframe containing all available Stan function defintions.
}
\description{
Available Stan Function Definitions
}
\examples{
available_stan_probability_distributions()
}
