% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{dpois_stan}
\alias{dpois_stan}
\title{Poisson Log Probability Mass Stan Code}
\usage{
dpois_stan()
}
\value{
A text string containing Stan code.
}
\description{
Stan has a built-in function which calculates the log probability density
of the the Poisson distribution:
\preformatted{
real poisson_lpmf(ints n | reals lambda)
}
}
