% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedPoissonLinked.R
\docType{data}
\name{zi_poisson_linked_lpmf_stan}
\alias{zi_poisson_linked_lpmf_stan}
\title{Zero-inflated Poisson Linked Log Probability Mass Function Stan Code}
\description{
Stan code for the zero-inflated Poisson distribution's log probability mass function,
where the amount of zero-inflation is a function of the mean.
}
\examples{
cat(zi_poisson_linked_lpmf_stan[["source_code"]])
}
\keyword{datasets}
