% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroInflatedPoisson.R
\docType{data}
\name{zi_poisson_lpmf_stan}
\alias{zi_poisson_lpmf_stan}
\title{Zero-inflated Poisson Log Probability Mass Function Stan Code}
\format{
An object of class \code{list} of length 1.
}
\usage{
zi_poisson_lpmf_stan
}
\value{
A string containing Stan source-code
}
\description{
Zero-inflated Poisson Log Probability Mass Function Stan Code
}
\examples{
cat(zi_poisson_lpmf_stan[["source_code"]])
}
\keyword{datasets}
